define seeders

+++ server/src/main/java/me/sajit/gym/workout_category/WorkoutCategoryService.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface WorkoutCategoryService{
     public List<WorkoutCategory> getAllWorkoutCategories();
-    WorkoutCategory getWorkoutCategoryById(int categoryId);
+    WorkoutCategory getWorkoutCategoryById(Long categoryId);
 }
